Retain precision in dividion
When dividing a value by an integer the value is rounded or
truncated. As the result is being stored in a float, it may
be more correct to retain the precision.

+++ app/src/main/java/videoeditor/bhuvnesh/com/ffmpegvideoeditor/views/VisualizerView.java
@@ -53,13 +53,13 @@ public class VisualizerView extends View {
         }
         mRect.set(0, 0, getWidth(), getHeight());
         for (int i = 0; i < mBytes.length - 1; i++) {
-            mPoints[i * 4] = mRect.width() * i / (mBytes.length - 1);
-            mPoints[i * 4 + 1] = mRect.height() / 2
-             + ((byte) (mBytes[i] + 128)) * (mRect.height() / 2) / 128;
-            mPoints[i * 4 + 2] = mRect.width() * (i + 1) / (mBytes.length - 1);
-            mPoints[i * 4 + 3] = mRect.height() / 2
-             + ((byte) (mBytes[i + 1] + 128)) * (mRect.height() / 2)
-             / 128;
+            mPoints[i * 4] = mRect.width() * i / (float)(mBytes.length - 1);
+            mPoints[i * 4 + 1] = mRect.height() / 2.0f
+             + ((byte) (mBytes[i] + 128)) * (mRect.height() / 2.0f) / 128.0f;
+            mPoints[i * 4 + 2] = mRect.width() * (i + 1) / (float)(mBytes.length - 1);
+            mPoints[i * 4 + 3] = mRect.height() / 2.0f
+             + ((byte) (mBytes[i + 1] + 128)) * (mRect.height() / 2.0f)
+             / 128.0f;
         }
         canvas.drawLines(mPoints, mForePaint);
     }
